% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hilama.R
\name{cal_eval_metric}
\alias{cal_eval_metric}
\title{Calculate the FDR and Power of the test given the true signal.}
\usage{
cal_eval_metric(Theta_true, Theta_hat, pvalue_mat, sig_level = 0.05)
}
\arguments{
\item{Theta_true}{True value of signal matrix.}

\item{Theta_hat}{Estimated value of signal matrix.}

\item{pvalue_mat}{Estimated p-value of Theta_true.}

\item{sig_level}{The significance level.}
}
\value{
A vector including estimated FDR, Power and mean bias over nonzero locations.
}
\description{
Calculate the FDR and Power of the test given the true signal.
}
