% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ddlasso.R
\name{esti_score}
\alias{esti_score}
\title{Estimate the score vector of a data matrix corresponding to the given column \code{id}.}
\usage{
esti_score(X, id, F_dc)
}
\arguments{
\item{X}{A matrix data.}

\item{id}{Column number of \code{X}.}

\item{F_dc}{The decorrelating matrix used to transform on the matrix data \code{X}.}
}
\value{
A score vector.
}
\description{

}
