% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ddlasso.R
\name{esti_score_all}
\alias{esti_score_all}
\title{Estimate the score matrix for a given data matrix.}
\usage{
esti_score_all(X, rho = 0.5, is.parallel = F, core_num = 4, standardize = TRUE)
}
\arguments{
\item{X}{A matrix data.}

\item{rho}{A parameter in (0,1) used to define the Trim transformation matrix.}

\item{is.parallel}{A logical value indicating whether to use parallel computation, default is TRUE.}

\item{core_num}{The number of CPU cores to use in parallel computation, default is depends on your system.}

\item{standardize}{A logical value indicating whether to standardize the data. Default is TRUE.}
}
\value{
A list of length \code{ncol(X)}, with each element returned from \code{esti_score()} function.
}
\description{
Estimate the score matrix of a data matrix using node-wise Lasso on the transformed data matrix.
}
