% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hilama.R
\name{get_sig_nie_nde}
\alias{get_sig_nie_nde}
\title{Return the significant NIE and NDE.}
\usage{
get_sig_nie_nde(result_hilama_ls, ratio_screen = 0.1, alpha = 0.1)
}
\arguments{
\item{result_hilama_ls}{Object returned by hilama_fast.}

\item{ratio_screen}{Number of pairs used for multiple testing. When the correlation among exposures is relatively high and the sample size is small, we recommend a high screening ratio to achieve greater power. In other scenarios, given the relatively minor changes in power, a moderate screening ratio may be considered to control the FDR.}

\item{alpha}{The nominal False Discovery Rate level.}
}
\value{
\item{result_nie}{data frame of significant NIE using JST.}
\item{result_nde}{data frame of significant NDE.}
}
\description{
Return the significant NIE and NDE.
}
