% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hilama.R
\name{hilama}
\alias{hilama}
\title{A function to identify the nonzero active paths from high dimensional exposures, high dimensional mediators and a continuous outcome
 while guaranteeing false discovery rate (FDR) control in finite sample settings.}
\usage{
hilama(X, M_multi, Y, is.parallel = T, core_num = 4, verbose = T, is.adap = F)
}
\arguments{
\item{X}{A matrix of input exposure data.}

\item{M_multi}{A matrix of input mediator data.}

\item{Y}{A vector of response data.}

\item{is.parallel}{A logical value indicating whether to use parallel computation, default is TRUE.}

\item{core_num}{The number of CPU cores to use in parallel computation, default is depends on your system.}

\item{verbose}{A logical value indicating whether to display progress info during computation, default is TRUE.}

\item{is.adap}{A logical value. If FALSE (default), the initial estimate in \code{infer_ddlasso()} uses LASSO. Otherwise, adaptive LASSO is used.}
}
\value{
A list containing \code{nie_mat_hat}, \code{pvalue_Theta}, \code{pvalue_beta}, \code{nde_vec_hat},
and \code{pvalue_nde}, as well as \code{coef_ls}, which contains the intermediate calculation information.
}
\description{
This function provides false discovery rate (FDR) control and support recovery in high-dimensional mediation models with
continuous outcome, where the exposure matrix, mediator matrix are potentially high-dimensional.
}
