% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ddlasso.R
\name{infer_ddlasso}
\alias{infer_ddlasso}
\title{Infer the coefficient high dimensional linear model with latent confounder using double debiased strategy.}
\usage{
infer_ddlasso(
  X,
  Y,
  rho = 0.5,
  is.parallel = F,
  core_num = 4,
  is.adap = T,
  score_ls = NULL,
  standardize = TRUE
)
}
\arguments{
\item{X}{A matrix of predictor data.}

\item{Y}{A vector of response data.}

\item{rho}{A scalar correlation parameter between the residual of X and Y given the confounder Z.}

\item{is.parallel}{A logical value indicating whether to use parallel computation, default is TRUE.}

\item{core_num}{The number of CPU cores to use in parallel computation, default is depends on your system.}

\item{is.adap}{A logical value, if FALSE (default), the initial estimate in \code{infer_ddlasso()} uses LASSO, otherwise, adaptive LASSO is used.}

\item{score_ls}{A list of score vectors, generated by \code{esti_score_all()}.}

\item{standardize}{A logical value indicating whether to standardize the data. Default is TRUE.}
}
\value{
A data frame with 5 columns. Column \code{init_coef} represents the initial estimates. Column \code{dd_coef} represents the double debiased estimate. Column \code{std}
represents the standard deviation of the estimates. Column \code{z_stat} represents the scaled estimates. Column \code{pvalue} represents the p-value of the estimates.
}
\description{
This function estimates and infers the coefficients of a high-dimensional linear model with a latent confounder using the Decorrelate & Debiase method proposed by Sun, Ma, and Xia (2022).
}
\references{
Sun, Y., Ma, L., & Xia, Y. (2022). A decorrelating and debiasing approach to simultaneous inference for high-dimensional confounded models.
arXiv preprint arXiv:2202.05162.
}
