% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hilama.R
\name{esti_score_all}
\alias{esti_score_all}
\title{Estimate the score matrix for a given data matrix.}
\usage{
esti_score_all(
  X,
  nlambda = 100,
  ratio_singular_min = 1.5,
  Khat = NULL,
  parallel = F,
  core_num = 4
)
}
\arguments{
\item{X}{A matrix data.}

\item{nlambda}{Number of hyper-parameter to search.}

\item{ratio_singular_min}{Used to check the existence of hidden confounder. If \eqn{\frac{\lambda_k}{\lambda_{k+1}}} is always smaller than this value,
then no hidden confounder exists and only single debiased procedure is applied.}

\item{Khat}{The number of possible hidden confounders.}

\item{parallel}{A logical value indicating whether to use parallel computation, default is TRUE.}

\item{core_num}{The number of CPU cores to use in parallel computation, default is depends on your system.}
}
\description{
Estimate the score matrix of a data matrix using node-wise Lasso on the transformed data matrix.
}
